requirements.txt
pymongo
flask
flask_restplus
marshmallow_jsonapi
flask_cors
python-dotenv